Fix partial completion variable calls
